feat: update folder and build script
