+++ packages/web-integration/modern.config.ts
@@ -7,6 +7,8 @@ export default defineConfig({
     format: 'cjs',
     input: {
       index: 'src/index.ts',
+      'bridge-mode': 'src/bridge-mode/index.ts',
+      'bridge-mode-browser': 'src/bridge-mode/browser.ts',
       utils: 'src/common/utils.ts',
       'ui-utils': 'src/common/ui-utils.ts',
       debug: 'src/debug/index.ts',
