+++ packages/web-integration/src/bridge-mode/browser.ts
@@ -1,0 +1,3 @@
+import { ChromeExtensionPageBrowserSide } from '../bridge-mode/page-browser-side';
+
+export { ChromeExtensionPageBrowserSide };
