+++ packages/web-integration/src/bridge-mode/common.ts
